## Description

Buffers writes for trace, memory file and output printing. Running with hyperfine gives the following results (used factorial_multirun benchmark for testing):
```
# Before buffering, no writes
Benchmark 1: ../target/release/cleopatra-run ../cairo_programs/benchmarks/factorial_multirun.json 
  Time (mean ± σ):     452.4 ms ±   7.9 ms    [User: 433.5 ms, System: 18.6 ms]
  Range (min … max):   443.7 ms … 468.0 ms    10 runs
# After buffering, no writes
Benchmark 1: ../target/release/cleopatra-run ../cairo_programs/benchmarks/factorial_multirun.json
  Time (mean ± σ):     458.0 ms ±   5.8 ms    [User: 443.1 ms, System: 14.6 ms]
  Range (min … max):   449.8 ms … 469.4 ms    10 runs

# Before buffering, writing trace
Benchmark 1: ../target/release/cleopatra-run ../cairo_programs/benchmarks/factorial_multirun.json --trace_file trace
  Time (mean ± σ):      2.275 s ±  0.021 s    [User: 0.694 s, System: 1.579 s]
  Range (min … max):    2.230 s …  2.308 s    10 runs
# After buffering, writing trace
Benchmark 1: ../target/release/cleopatra-run ../cairo_programs/benchmarks/factorial_multirun.json --trace_file trace
  Time (mean ± σ):     517.4 ms ±   5.3 ms    [User: 472.5 ms, System: 44.4 ms]
  Range (min … max):   508.7 ms … 526.9 ms    10 runs

# Before buffering, writing memory
Benchmark 1: ../target/release/cleopatra-run ../cairo_programs/benchmarks/factorial_multirun.json --memory_file memory
  Time (mean ± σ):     525.7 ms ±  10.0 ms    [User: 497.0 ms, System: 28.2 ms]
  Range (min … max):   514.8 ms … 548.6 ms    10 runs
# After buffering, writing memory
Benchmark 1: ../target/release/cleopatra-run ../cairo_programs/benchmarks/factorial_multirun.json --memory_file memory
  Time (mean ± σ):     526.6 ms ±  15.4 ms    [User: 495.3 ms, System: 30.7 ms]
  Range (min … max):   516.5 ms … 568.4 ms    10 runs

# Before buffering, printing output
Benchmark 1: ../target/release/cleopatra-run ../cairo_programs/benchmarks/factorial_multirun.json --print_output
  Time (mean ± σ):     466.3 ms ±  11.4 ms    [User: 447.3 ms, System: 18.7 ms]
  Range (min … max):   458.6 ms … 496.8 ms    10 runs
# After buffering, printing output
Benchmark 1: ../target/release/cleopatra-run ../cairo_programs/benchmarks/factorial_multirun.json --print_output
  Time (mean ± σ):     457.1 ms ±   2.6 ms    [User: 442.7 ms, System: 14.1 ms]
  Range (min … max):   453.5 ms … 460.6 ms    10 runs
```

There's a very significant improvement in writing trace files.